refactor(post_gen): Added file check before removal
Added a file exists check in the remove_file function so that no errors will happen if the file to delete does not actually exist.

+++ hooks/post_gen_project.py
@@ -8,6 +8,8 @@ PROJECT_DIRECTORY = os.path.realpath(os.path.curdir)
 
 
 def remove_file(filepath):
+    if not Path(filepath).exists:
+        return
     os.remove(os.path.join(PROJECT_DIRECTORY, filepath))
